refactor(mongo_adapter): extract trait_impl (cycle34)
Split mod.rs (324 LOC) by moving the DatabaseInterface trait impl into
trait_impl.rs. Pure code re-arrangement; behavior unchanged.

- mod.rs: 41 LOC (was 324)
- trait_impl.rs: 289 LOC (new)
